Moving all role related things to roles.
